heed

Module iteration_method

The set of possible iteration methods for the different iterators.

Enums

MoveBetweenKeys

Moves between keys and ignores the duplicate values of keys.

MoveOnCurrentKeyDuplicates

Moves only on the duplicate values of a given key and ignores other keys.

MoveThroughDuplicateValues

Moves to the next or previous key if there are no more values associated with the current key.

Traits

IterationMethod

The trait used to define the way iterators behave.