heed::iteration_method

Trait IterationMethod

pub trait IterationMethod {
    const MOVE_OPERATION: MoveOperation;
}

The trait used to define the way iterators behave.

Required Associated Constants

const MOVE_OPERATION: MoveOperation

The internal operation to move the cursor through entries.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.